chore: rename the package_registry_info module to package_registry_spec
