+++ src/macaron/slsa_analyzer/checks/provenance_witness_l1_check.py
@@ -22,7 +22,7 @@ from macaron.slsa_analyzer.provenance.witness import (
 )
 from macaron.slsa_analyzer.registry import registry
 from macaron.slsa_analyzer.slsa_req import ReqName
-from macaron.slsa_analyzer.specs.package_registry_info import PackageRegistryInfo
+from macaron.slsa_analyzer.specs.package_registry_spec import PackageRegistryInfo
 
 logger: logging.Logger = logging.getLogger(__name__)
